Add files via upload
# Evaluation Forms for 42 Project Cursus
This repository contains the evaluation forms for all 42 projects in the cursus. The forms are designed to assess the performance of students on each project and provide feedback for improvement. The forms are organized by project name and rank. Currently, the forms are available in French only, but English versions will be added as soon as possible.

# Contributing
I welcome contributions to this repository. If you'd like to add an evaluation form for a project or improve an existing form, please follow these steps:

1. Fork the repository
2. Clone your fork to your local machine
3. Update a correction or add one
4. Commit the changes with a descriptive message
5. Push the changes to your fork
6. Submit a pull request

# Using the Forms
The forms can be used by students to assess the performance of each of their project. They provide a standardized way of evaluating the projects and provide a basis for constructive feedback. To use the forms, simply fill in the sections for comments and grading.

# Language
The forms are in french for now but I will post english version as soon as possible.

### Evaluation Status
The following table shows the status of each evaluation form.

<div align="center">

| Project                                      | Status                              |
| -------------------------------------------- | ----------------------------------- |
| [Libft](./Rank00/Libft)                      | Added !                             |
| [ft_printf](./Rank01/ft_printf)              | Not yet added                       |
| [Get_next_line](./Rank01/get_next_line)      | Not yet added                       |
| [Born2beroot](./Rank01/Born2beroot)          | Added !                             |
| [push_swap](./Rank02/push_swap)              | Added !                             |
| [so_long](./Rank02/42-so_long)               | Not yet added                       |
| [fract-ol](./Rank02/fract-ol)                | Added !                             |
| [FdF](./Rank02/FdF)                          | Added !                             |
| [pipex](./Rank02/pipex)                      | Added !                             |
| [minitalk](./Rank02/minitalk)                | Added !                             |
| [minishell](./Rank03/minishell)              | Added !                             |
| [Philosophers](./Rank03/Philosophers)        | Added !                             |
| [netpractice](./Rank04/netpractice)          | Added !                             |
| [CPP](./Rank04/CPP_Modules/)                 | Added but 5,6 and 8 still missing ! |
| [cub3d](./Rank04/cub3d)                      | Added !                             |
| [miniRT](./Rank04/miniRT)                    | Not yet added                       |
| [Inception](./Rank05/Inception)              | Not yet added                       |
| [ft_containers](./Rank05/ft_containers)      | Not yet added                       |
| [webserv](./Rank05/webserv)                  | Not yet added                       |
| [ft_irc](./Rank05/ft_irc)                    | Not yet added                       |
| [ft_transcendence](./Rank06/ft_transcendence)| Not yet added                       |

</div>

# Feedback
If you have any feedback or suggestions for improvement, please open an issue or submit a pull request. I'd love to hear from you!